chore: Move `ModifyLyricBase`
- I think instead of the global `sheets` folder, we'll probably have localized `sheet` folders in each screen subfolder for the sheets that go in any of those screens.
