+++ mobile/src/navigation/screens/lyrics/components/ModifyViewBase.tsx
